test(measurement_retrieval): move collect_epoch_measurements tests to the right file
Tests for collect_epoch_measurements belong in measurement_retrieval.rs
alongside the function they test, not in audit.rs.
