+++ git_perf/src/measurement_retrieval.rs
@@ -134,6 +134,121 @@ where
 ///     println!("Commit: {}", commit.commit);
 /// }
 /// ```
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::data::{Commit, MeasurementData};
+    use crate::stats::ReductionFunc;
+    use anyhow::Result;
+
+    fn make_commit_with_measurement(sha: &str, name: &str, val: f64) -> Commit {
+        Commit {
+            commit: sha.to_string(),
+            title: String::new(),
+            author: String::new(),
+            measurements: vec![MeasurementData {
+                epoch: 1,
+                name: name.to_string(),
+                timestamp: 0.0,
+                val,
+                key_values: std::collections::HashMap::new(),
+            }],
+        }
+    }
+
+    fn make_commits_iter(commits: &[Result<Commit>]) -> impl Iterator<Item = Result<Commit>> + '_ {
+        commits.iter().map(|r| match r {
+            Ok(c) => Ok(c.clone()),
+            Err(e) => Err(anyhow::anyhow!("{}", e)),
+        })
+    }
+
+    #[test]
+    fn test_collect_epoch_measurements_basic() {
+        // 5 commits with values 10.0..50.0, same epoch
+        let commits: Vec<Result<Commit>> = (0..5)
+            .map(|i| {
+                Ok(make_commit_with_measurement(
+                    &format!("sha{:040}", i),
+                    "bench",
+                    (i as f64 + 1.0) * 10.0,
+                ))
+            })
+            .collect();
+
+        let (values, shas) = collect_epoch_measurements(
+            make_commits_iter(&commits),
+            |m: &MeasurementData| m.name == "bench",
+            &ReductionFunc::Min,
+        );
+        assert_eq!(values.len(), 5);
+        assert_eq!(shas.len(), 5);
+        assert!((values[0] - 10.0).abs() < f64::EPSILON); // HEAD is newest (index 0)
+    }
+
+    #[test]
+    fn test_collect_epoch_measurements_filters_by_name() {
+        let commits: Vec<Result<Commit>> =
+            vec![Ok(make_commit_with_measurement("sha1", "other", 99.0))];
+        let (values, shas) = collect_epoch_measurements(
+            make_commits_iter(&commits),
+            |m: &MeasurementData| m.name == "bench",
+            &ReductionFunc::Min,
+        );
+        assert!(values.is_empty());
+        assert!(shas.is_empty());
+    }
+
+    #[test]
+    fn test_collect_epoch_measurements_stops_at_epoch_change() {
+        let mut commits: Vec<Result<Commit>> = Vec::new();
+        // First 3 commits: epoch 1
+        for i in 0..3 {
+            commits.push(Ok(Commit {
+                commit: format!("sha{}", i),
+                title: String::new(),
+                author: String::new(),
+                measurements: vec![MeasurementData {
+                    epoch: 1,
+                    name: "bench".to_string(),
+                    timestamp: 0.0,
+                    val: 10.0,
+                    key_values: std::collections::HashMap::new(),
+                }],
+            }));
+        }
+        // Next 2 commits: epoch 2 (older, different epoch)
+        for i in 3..5 {
+            commits.push(Ok(Commit {
+                commit: format!("sha{}", i),
+                title: String::new(),
+                author: String::new(),
+                measurements: vec![MeasurementData {
+                    epoch: 2,
+                    name: "bench".to_string(),
+                    timestamp: 0.0,
+                    val: 20.0,
+                    key_values: std::collections::HashMap::new(),
+                }],
+            }));
+        }
+
+        let (values, _) = collect_epoch_measurements(
+            make_commits_iter(&commits),
+            |m: &MeasurementData| m.name == "bench",
+            &ReductionFunc::Min,
+        );
+        // Should only return epoch 1 data (3 commits)
+        assert_eq!(values.len(), 3);
+        for v in &values {
+            assert!(
+                (v - 10.0).abs() < f64::EPSILON,
+                "Should only return epoch 1 values"
+            );
+        }
+    }
+}
+
 pub fn walk_commits_from(
     start_commit: &str,
     num_commits: usize,
